+++ keys.nix
@@ -8,6 +8,6 @@
 
   hosts = {
     phi = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIMxOi/S1TLBg8/ZRX5XfCTlM8A+I0q0pQksrxtfjdYFP";
-    sigma = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIN3pv/iYVewBlpTImGMvChu7da6jsi99/Y0vpsmAoeds";
+    sigma = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIDxRoznXzz/T6s5UeHG1uoHCXGfXSpy27eTEzC0/EUW+";
   };
 }